Made some changes to shader, and completed testing coverage on shader. Created a base test file to handle the python version stuff.
